chore(ui): fix false rename
